Re-wrote script to accomodate both developers and anonymous pserver clients.
I've set it up so that if you set the environment variable SF_USERNAME to be, well, your sourceforge username, it will log you in as a developer and grab the stuff.

If there is no env var set then it asks if you are a dev, if you say no then it does the anonymous pserver thing, if yes it ask for your sf username and sets the right vars.

I hope you don't mind but I swapped it to perl because I couldn't get the if statements in csh to work right (and csh has very bad error reporting).  I hope noone minds. Since we're on OS X we can be sure that perl is installed.

So that's my big contribution for the past three months!


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@326 0fceea05-a30d-0410-8a8b-80ef821fb0a1