Change to be consistent with the plist, which wants RIS/Medline instead of PubMed; this allows PubMed saving, although we have errors.  There are some problems with writeToFile and/or dataRepresentationOfType in this class, I think.
1)  Relocating an RIS/Medline file to another location on
disk with Save or Save As is not allowed, since writeToFile changes the
filename extension and writes it in the same directory as the original file.
Is this consistent with the documentation on overrides of writeToFile in
the NSDocument class, regarding assumptions about the file?

2)  Every save operation (even a successful bibtex save) of a .fcgi file
results in an error sheet saying "Couldn't save document."  See bug
897765 which describes this issue in detail.

I think it might be better to allow BibDesk to be an editor of RIS/Medline,
then force the user to Save As bibtex, although I can't get this to work
perfectly either with "return [super saveDocumentAs:self]" in writeToFile.
If the user chose to overwrite their .fcgi file in this case, the Save As op
would remove the file, then fail to write it out.


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@432 0fceea05-a30d-0410-8a8b-80ef821fb0a1