Partial fix for bug number 880368.
Multiple Authors are now saved, and we now use FAU instead of AU to get their most detailed names. I'm not totally sure this is always available though.
If it's not always there we will need to try to prefer FAU over AU but use AU if it's the only one, which might get hairy.


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@479 0fceea05-a30d-0410-8a8b-80ef821fb0a1