Fix for  897765, PubMed .fcgi Saving Problem.
We are a viewer for PubMed files, but we convert them to bibtex.  Having
converted them, we need to change the fileType and fileName to
correspond to bibtex, or else the Save operation gives an error.
Previously, only saving to the same directory was allowed for PubMed files,
and the user was not warned if an existing file was overwritten.
A unique name is now assigned based on the imported filename, and the
user has the option to relocate the file or change the name using Save As.

The override of writeToFile:fileName ofType:docType in BibDocument is
now unnecessary, but remains in the code.


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@486 0fceea05-a30d-0410-8a8b-80ef821fb0a1