Add a couple of characters to conversions list.
The additions caused [astring cString] to puke ungracefully and refuse to open my file.  To reproduce, make a test.bib with this entry in it:

@article{liu1991,
	Author = {Dobroslav Znidar{\u c}i{\'c}}}

I added a check for cString conversion, and fallback to lossyCString.
The display will be fouled up in the authors table, but that is the only
problem that I can see.


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@535 0fceea05-a30d-0410-8a8b-80ef821fb0a1