Changes to view local file button - now it's a popup with a down arrow.
For now the only two menu items are view (which has to be at the top)
and choose file.

So you can change the file from the button using the down arrow even if there's a file already chosen.

The default icon for 'no file' is from the system headers, it looks like crap.
Need to find a better image for the disabled button.

Other UI tweaks need to be done as well.

Finally, the menu can be customized for some smartness, like to show unclaimed files in the papers folder or recent downloads from safari.


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@710 0fceea05-a30d-0410-8a8b-80ef821fb0a1