More changes to the file icon button.
The no files icon is now Martin Kerz's much nicer grey file icon with
question mark.

The menu is built dynamically.

There is a 'reveal in Finder' option.

There's a set of menu items for Safari recent downloads, and for
recently viewed files in Preview.

I attempted to make the menu update as Preview updates, but it seems
to wait for the program to quit, even though I tell it to synchronize
Preview's prefs using the CFPreferences low level api...


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@725 0fceea05-a30d-0410-8a8b-80ef821fb0a1