Changes to allow undo of adding and removing fields
Changes to set modified dates on items you edit and created dates on new ones.

Modified dates behave as you'd like through undo/redo.
 This was kind of tricky but it works without changing external interfaces.

However, undo's change count is a little flaky.  It seems to err on
the side of asking you to save when there are no changes, so I can
deal with that for now.

Added display for modified/created dates: these can be shown in
columns with titles "Modified" and "Created". Probably we should also
display them elsewhere.  WIthout the columns you can see them in the
preview.

To avoid a warning, I changed BibItem setFields to take an
NSDictionary argument. this is less restrictive, but just fine since
it makes a mutable copy anyway.


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@794 0fceea05-a30d-0410-8a8b-80ef821fb0a1