Fixed two errors that prevented sorting by the modification and creation dates from working:
-> in BibDocument.m the wrong names for the table columns were used when deciding which comparison method to use

-> in BibItem.m the sort methods for the creation and modification dates still used the publication date in one part (probably left over from copying the method for sorting by publication date)

Because I found the original comparison methods hard to understand, I rewrote it more explicitly (i.e. the changes look bigger than they are) and now the real work is done by a single method that gets called by the others with the appropriate values to avoid having duplicate code.


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@1373 0fceea05-a30d-0410-8a8b-80ef821fb0a1