Fix bug #1198028, editor forgets continuous spell checking setting.
This fix is not perfect, as noted in the implementation; in particular, the preference can only be changed if the window itself is first responder.

It would probably be possible to fix this by using a custom NSTextView subclass for the window's field editor and the text views in the tabs, but I'm not convinced it's worth the effort (I tried it and lost tab behavior in the form, for instance).


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@2281 0fceea05-a30d-0410-8a8b-80ef821fb0a1