Continuing fix for bug #1199338.  Pasted publications do not have a unique file order until they are saved, and their final order in the document is undefined.
We no longer sort the publications array, so its order should be file order; shownPublications is used for display and sorting, and this also saves overhead.  Publications are now added to the end of the unsorted publications array, so ordering should no longer be random (actually, I think new entries were assigned integers starting from 1, which explains why new entries were being added at the beginning of a file).

Current sort order is now  maintained during a search; alternately, we should de-sort the table when searching (if this turns out to be too slow).


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@2293 0fceea05-a30d-0410-8a8b-80ef821fb0a1