Declare cleanup as oneway, so we don't block and possibly wait for port timeouts when disabling browsing.
Copy the shared group name, since it doesn't change over the lifetime of either object, and since we can't guarantee that the shared group will be around for the lifetime of the server (but the name may be required for authentication).  This is mandatory for cleanup to be oneway.

Check flag before messaging the group that it needs an update, in case that message comes from the remote server during cleanup (when the group is garbage).

Add more comments, so we don't forget what these variables and methods are for (remote, local, main may not make sense next year).

git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@6258 0fceea05-a30d-0410-8a8b-80ef821fb0a1