Fix actions for toolbar items in text only mode.
Make sure an item with crossref is created when the option key is hold down for new  publciation. This fixes at least the action for the New toolbar item in text only mode, the label is still not updated. 

Allow selecting the quick search field from the toolbar in text only mode. 

git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@6904 0fceea05-a30d-0410-8a8b-80ef821fb0a1