Don't register individual BibItems for notifications; use the document to observe changes in custom fields and types, and then apply these to their items.
Profiling with Shark showed that 40% of the overhead from sharing a large number of publications was due to removing each item as a notification observer in its dealloc method (after copy/release).  This also makes closing a large document faster, as the same problem was evident there.

An implication of this change is that shared items no longer receive these notifications, since they don't have a document.  However, custom fields and types are irrelevant in this case, as these items can't be user-edited.

git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@7245 0fceea05-a30d-0410-8a8b-80ef821fb0a1