Fix for bug # 1596267.
Convert data to UTF-8 string encoding before parsing BibTeX data with encodings for which btparse has problems. 

Known encodings with problems are Unicode and Japanese (Shift-JIS), more might be added later. 

Also need to save and read the group data in the document encoding for those encodings, or we cannot convert the file data. 

git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@8606 0fceea05-a30d-0410-8a8b-80ef821fb0a1