Database find & replace redesign.
Place the Replace All options and Find options in boxes, so the title gioves more info about their function. 

Replace replace all scope popup by a matrix, so it is more obvious that the scope can be the shown items as opposed to the whole database.

Add a Wrap Around option. 

 Fix enabling of search scope popup. 

git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@9338 0fceea05-a30d-0410-8a8b-80ef821fb0a1