BibItem and BibEditor were both observing custom fields change notification and calling makeType.  However, more cached values in the BibItem need to be flushed, since field types can change.  The editor also needed to reload the author table.
Have the editor ensure that its publication receives the notification first, since we don't know if editors or the document are notified first.  This removes the ordering dependency.

Since the group field and its type can change, rebuild groups in the document and update the previews when custom fields have been changed.  This fixes the exception that Alex reported to the develop list (changing current group textual field's type to person field).

git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@10353 0fceea05-a30d-0410-8a8b-80ef821fb0a1