Revise searching API to use SearchKit, maintaining separate indexes for three term collections (any field, title, and author).  Other indexes can be added.  There is a single new class to manage the search button view.  Code for adding and removing that view is still in a semi-finished state.
BDSKLevelIndicatorCell is now in a separate file, since it's useful
in other parts of the codebase that shouldn't import FileMatcher.

The old substring-based search API has been minimalized and left in
place for Services compatibility, in case anyone still uses it.

Preferences aren't solidified yet, and neither is the interaction with
external/search groups (do we need to search those?).


git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@10396 0fceea05-a30d-0410-8a8b-80ef821fb0a1