Add some pub type heuristics instead of always using article.
Use all paragraphs of the abstract instead of just the first one.

Add a fallback for cases without AuCollectiveName (older database entries), which fixes a lot of cases.

Request some additional fields from the service.  It appears that article_no is a DOI (or some other library identifier), so use DOI instead of number.

git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@10817 0fceea05-a30d-0410-8a8b-80ef821fb0a1