Unify some temporary file and directory creation code by moving it all into the NSFileManager category.  Separate methods by directory creation (file created by manager) and file name creation (file created by caller).  Hopefully the names are clear enough.
The temporary directory creation needs to be thread safe, and guarded against race conditions, so try to ensure this.

The temporary file creation methods are used only on the main thread, so at least aren't subject to a race between threads.  Security implications aren't significant enough for me to worry about at this point.

git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@10949 0fceea05-a30d-0410-8a8b-80ef821fb0a1